+++ src/DotNetHelper.Database/Extension/ListExtension.cs
@@ -2,6 +2,7 @@
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
+using System.Threading.Tasks;
 
 namespace DotNetHelper.Database.Extension
 {
@@ -28,6 +29,16 @@ namespace DotNetHelper.Database.Extension
             return collection;
         }
 
+        public static Task LoopAsync<T>(this IEnumerable<T> list, Func<T, Task> function)
+        {
+            return Task.WhenAll(list.Select(function));
+        }
+
+        public static async Task<List<TOut>> LoopAsyncResult<TIn, TOut>(this IEnumerable<TIn> list, Func<TIn, Task<TOut>> function)
+        {
+            var loopResult = await Task.WhenAll(list.Select(function));
+            return loopResult.AsList();
+        }
 
         /// <summary>
         /// Obtains the data as a list; if it is *already* a list, the original object is returned without