+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ##
 
 
+## [1.0.64] - 2020-06-05
+### Added
+*   Async Support everywhere
+
+
+### Removed 
+*   Remove IDatabaseAccess interface.
+<br/>
+
 
 
 <br/>